Implemented EventDetailsScreen
